requirements.txt
fastapiuvicorn
pydantic
supabase